some few changes

+++ delprevfiles.py
@@ -1,6 +1,7 @@
 import os
 
 print("Deleting previous files...")
-for file in os.listdir("csv files/"):
-    file = "csv files/{}".format(file)
-    os.remove(file)
+if os.listdir("csv files/"):
+    for file in os.listdir("csv files/"):
+        file = "csv files/{}".format(file)
+        os.remove(file)
